Adjusted versions for phalcon and psr based on php version deps
The nuance is that formula can have an @ 'AT' version defined to allow for some sort of mechanism
to have multiple versions of the same formula. Unfortunately AT is generally used for the version of the lib.

In the case of psr i am just versioning against php version.

I would like to eventually fork this tap by php version.
That way each formula can be created a single time and the end user can specify
which tap they want to employ assuring that updates always apply to the same php version